:lipstick:: Changed QRScan to main page (additonally fixed minor visual features)
