+++ cypress/integration/app.spec.js
@@ -1,7 +1,7 @@
 describe('Form Validation', () => {
   it('should not submit the form if any field is empty', () => {
     // Start from the index page
-    cy.visit('http://localhost:3000/');
+    cy.visit('http://localhost:3000/uid');
     // Click on the form button
     cy.get('form').submit();
     // Check if the error message is displayed
