+++ styles/Home.module.css
@@ -54,7 +54,7 @@
 
 .translateIcon {
   position: fixed;
-  margin-bottom: 52rem;
+  margin-bottom: 55rem;
   margin-right: 13.5rem;
   z-index: 9;
 }
@@ -68,7 +68,7 @@
 .dropdown {
   position: fixed;
   width: 200px;
-  margin-bottom: 34.5rem;
+  margin-bottom: 37.5rem;
   margin-right: 4.5rem;
   z-index: 2;
   border: 1px solid rgba(0, 0, 0, 0.04);
